# Use bufio and don't sync FS