Found via profiling that the vast majority of the time taken during extraction was `write` syscalls.

This empirically makes things much faster on Kubernetes, n=1:

Before:
1481.40s user 996.60s system 282% cpu 14:36.62 total
After:
606.63s user 12.59s system 338% cpu 3:02.91 total

Was there a reason we were running `Sync` before? My understanding is that it's not  necessary unless we want trap files to be uncorrupted if the system crashes.

I've kicked off an evaluation.