@max-schaefer 2020-03-06
> the vast majority of the time taken during extraction was write syscalls

:tada:

> n=1

?

> Was there a reason we were running Sync before?

I don't think so.