@sauyon 2020-03-06
[Dist-compare](https://git.semmle.com/sauyon/dist-compare-reports/tree/8cf727432f03b7c71afbed36885f531dd579627b) (internal link) seems clean. It doesn't measure extraction time, right?

> > n=1
> 
> ?

Sorry, sample size of 1.