@max-schaefer 2020-03-06
> It doesn't measure extraction time, right?

Correct. The Go Differences job, on the other hand, does; I'll set off an experiment.

> sample size of 1

Thanks for the clarification.