@max-schaefer 2020-04-15
@sauyon, could you add a change note for this? I think it's user-visible enough to warrant one.