@max-schaefer 2020-04-15
Change note added in #101.